requirements.txt
alabaster==0.7.12
appdirs==1.4.4
attrs==19.3.0
Babel==2.8.0
certifi==2020.4.5.2
chardet==3.0.4
chord-lib==0.9.0
codecov==2.1.7
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
coverage==5.1
distlib==0.3.0
Django==2.2.13
django-filter==2.2.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-nose==1.4.6
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-camel-case==1.1.2
docutils==0.16
elasticsearch==7.6.0
entrypoints==0.3
fhirclient==3.2.0
filelock==3.0.12
flake8==3.8.3
idna==2.9
imagesize==1.2.0
importlib-metadata==1.6.0
isodate==0.6.0
itypes==1.2.0
Jinja2==2.11.2
jsonschema==3.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.4.0
nose==1.3.7
openapi-codec==1.3.2
packaging==20.4
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.8.2
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
rdflib==4.2.2
rdflib-jsonld==0.4.0
redis==3.5.3
requests==2.24.0
rfc3987==1.3.8
simplejson==3.17.0
six==1.15.0
snowballstemmer==2.0.0
Sphinx==2.4.4
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
strict-rfc3339==0.7
toml==0.10.1
tox==3.15.2
uritemplate==3.0.1
urllib3==1.25.9
virtualenv==20.0.23
Werkzeug==1.0.1
wincertstore==0.2
zipp==3.1.0
